Changes event date column from camelCase to snake_case and the respective updates on EventsController. Limits edition of user info to the user itself.

+++ app/Http/Controllers/UsersController.php
@@ -32,7 +32,7 @@ class UsersController extends Controller
     }
 
     public function getUser($identification)
-    {   
+    {
         $user = $this->fetchUser($identification);
 
         if($user == NULL){ //If it reaches here, user doesn't exist
@@ -59,8 +59,13 @@ class UsersController extends Controller
             if($user == NULL) {
                 return new CustomJsonResponse(false,"User not found", 404);
             }
-            $user->update($input);
-            return new CustomJsonResponse(true, $user, 200);
+            if($user->id == $request->user()->id) {
+                $user->update($input);
+                return new CustomJsonResponse(true, $user, 200);
+            } else {
+                return new CustomJsonResponse(false, "No permission to edit member", 403);
+            }
+
         } else {
             return new CustomJsonResponse(false, $validator->errors()->all(), 400); //400 is Bad Request
         }
@@ -114,13 +119,17 @@ class UsersController extends Controller
 
     public function deleteUser(Request $request, $identification)
     {
-        // TODO: Falta verificar as permissoes do user para editar este objeto (se admin ou utilizador)
-        $user = $this->getUser($identification);
+        $user = $this->fetchUser($identification);
         if($user == NULL) {
             return new CustomJsonResponse(false,"User not found", 404);
         }
-        $user->delete();
-        return new CustomJsonResponse(true, "User successfully deleted", 200);
+
+        if($user->id == $request->user()->id) {
+            $user->delete();
+            return new CustomJsonResponse(true, "User successfully deleted", 200);
+        }
+        return new CustomJsonResponse(false, "No permissions to delete member", 403);
+
     }
     /*
      * Valida as informações vindas do utilizador
@@ -145,7 +154,7 @@ class UsersController extends Controller
             $validationArray = [
                 'username' => 'unique:users',
                 'email' => 'email|unique:users',
-                'student_id' => ["unique:users", "digits:10"], 
+                'student_id' => ["unique:users", "digits:10"],
                 'password' => 'confirmed', // o "confirmed" necessita de um text input com o nome de password_confirmation, para verificar se a password é igual nos dois sítios.
 		        'course_id' => 'exists:courses,id',
                 'role_id' => 'exists:roles,id',
@@ -165,15 +174,20 @@ class UsersController extends Controller
         if($user == NULL){ //If it reaches here, user doesn't exist
             return new CustomJsonResponse(false,"User not found", 404);
         }
+        if($user->id == $request->user()->id) {
+            if(\Validator::make(['event_id' => $event_id], ['event_id' => 'exists:events,id'])->passes()) {
 
-        if(\Validator::make(['event_id' => $event_id], ['event_id' => 'exists:events,id'])->passes()) {
-            if(!$user->events->contains($event_id)) { //event already exists
-                $user->events()->attach($event_id);
-                $user->load('events'); //Refreshes the model
+                if(!$user->events->contains($event_id)) { //event already exists
+                    $user->events()->attach($event_id);
+                    $user->load('events'); //Refreshes the model
+                }
+                return new CustomJsonResponse(true, $user, 200);
             }
-            return new CustomJsonResponse(true, $user, 200);
+            return new CustomJsonResponse(false, "Event does not exist", 404);
+        } else {
+            return new CustomJsonResponse(false, "No permission to edit member info", 403);
         }
-        return new CustomJsonResponse(false, "Event does not exist", 404);
+
     }
 
     public function deleteEvent(Request $request, $identification, $event_id) {
@@ -183,14 +197,18 @@ class UsersController extends Controller
             return new CustomJsonResponse(false,"User not found", 404);
         }
 
-        foreach ($user->events as $event) {
-            if($event->id == $event_id) {
-                $user->events()->detach($event_id);
-                $user->load('events'); //Refreshes the model
-                return new CustomJsonResponse(true, $user, 200);
+        if($user->id == $request->user()->id) {
+            foreach ($user->events as $event) {
+                if ($event->id == $event_id) {
+                    $user->events()->detach($event_id);
+                    $user->load('events'); //Refreshes the model
+                    return new CustomJsonResponse(true, $user, 200);
+                }
             }
+            return new CustomJsonResponse(false, "User isn't participating in this event", 404);
+        } else {
+            return new CustomJsonResponse(false, "No permission to edit member info", 403);
         }
-        return new CustomJsonResponse(false, "User isn't participating in this event", 404);
     }
 
     public function postProject(Request $request, $identification, $project_id) {
@@ -200,14 +218,18 @@ class UsersController extends Controller
             return new CustomJsonResponse(false,"User not found", 404);
         }
 
-        if(\Validator::make(['project_id' => $project_id], ['project_id' => 'exists:projects,id'])->passes()) {
-            if(!$user->projects->contains($project_id)) { //project already exists
-                $user->projects()->attach($project_id);
-                $user->load('projects'); //Refreshes the model
+        if($user->id == $request->user()->id) {
+            if(\Validator::make(['project_id' => $project_id], ['project_id' => 'exists:projects,id'])->passes()) {
+                if (!$user->projects->contains($project_id)) { //project already exists
+                    $user->projects()->attach($project_id);
+                    $user->load('projects'); //Refreshes the model
+                }
+                return new CustomJsonResponse(true, $user, 200);
             }
-            return new CustomJsonResponse(true, $user, 200);
+            return new CustomJsonResponse(false, "Project does not exist", 404);
+        } else {
+            return new CustomJsonResponse(false, "No permission to edit member info", 403);
         }
-        return new CustomJsonResponse(false, "Project does not exist", 404);
     }
     
     public function deleteProject(Request $request, $identification, $project_id) {
@@ -217,14 +239,18 @@ class UsersController extends Controller
             return new CustomJsonResponse(false,"User not found", 404);
         }
 
-        foreach ($user->projects as $project) {
-            if($project->id == $project_id) {
-                $user->projects()->detach($project_id);
-                $user->load('projects'); //Refreshes the model
-                return new CustomJsonResponse(true, $user, 200);
+        if($user->id == $request->user()->id) {
+            foreach ($user->projects as $project) {
+                if($project->id == $project_id) {
+                    $user->projects()->detach($project_id);
+                    $user->load('projects'); //Refreshes the model
+                    return new CustomJsonResponse(true, $user, 200);
+                }
             }
+            return new CustomJsonResponse(false, "User isn't member of this project", 404);
+        } else {
+            return new CustomJsonResponse(false, "No permission to edit member info", 403);
         }
-        return new CustomJsonResponse(false, "User isn't member of this project", 404);
     }
 
     public function postOrganization(Request $request, $identification, $organization_id) {
@@ -234,31 +260,41 @@ class UsersController extends Controller
             return new CustomJsonResponse(false,"User not found", 404);
         }
 
-        if(\Validator::make(['organization_id' => $organization_id], ['organization_id' => 'exists:organizations,id'])->passes()) {
-            if(!$user->organizations->contains($organization_id)) {
-                $user->organizations()->attach($organization_id);
-                $user->load('organizations'); //Refreshes the model
-            }
-            return new CustomJsonResponse(true, $user, 200);
-        }
-        return new CustomJsonResponse(false, "Organization does not exist", 404);
-    }
-
-    public function deleteOrganization(Request $request, $identification, $organization_id) {
-        $user = $this->fetchUser($identification);
-
-        if($user == NULL){ //If it reaches here, user doesn't exist
-            return new CustomJsonResponse(false,"User not found", 404);
-        }
-
-        foreach ($user->organizations as $organization) {
-            if($organization->id == $organization_id) {
-                $user->organizations()->detach($organization_id);
-                $user->load('organizations'); //Refreshes the model
+        if($user->id == $request->user()->id) {
+            if(\Validator::make(['organization_id' => $organization_id], ['organization_id' => 'exists:organizations,id'])->passes()) {
+                if(!$user->organizations->contains($organization_id)) {
+                    $user->organizations()->attach($organization_id);
+                    $user->load('organizations'); //Refreshes the model
+                }
                 return new CustomJsonResponse(true, $user, 200);
             }
+            return new CustomJsonResponse(false, "Organization does not exist", 404);
+        } else {
+            return new CustomJsonResponse(false, "No permission to edit member info", 403);
         }
-        return new CustomJsonResponse(false, "User isn't member of this organization", 404);
+
+    }
+
+    public function deleteOrganization(Request $request, $identification, $organization_id)
+    {
+        $user = $this->fetchUser($identification);
+
+        if ($user == NULL) { //If it reaches here, user doesn't exist
+            return new CustomJsonResponse(false, "User not found", 404);
+        }
+
+        if ($user->id == $request->user()->id) {
+            foreach ($user->organizations as $organization) {
+                if ($organization->id == $organization_id) {
+                    $user->organizations()->detach($organization_id);
+                    $user->load('organizations'); //Refreshes the model
+                    return new CustomJsonResponse(true, $user, 200);
+                }
+            }
+            return new CustomJsonResponse(false, "User isn't member of this organization", 404);
+        } else {
+            return new CustomJsonResponse(false, "No permission to edit member info", 403);
+        }
     }
 
     public function postProgrammingLanguage(Request $request, $identification, $programming_language_id) {
@@ -268,14 +304,19 @@ class UsersController extends Controller
             return new CustomJsonResponse(false,"User not found", 404);
         }
 
-        if(\Validator::make(['programming_language_id' => $programming_language_id], ['programming_language_id' => 'exists:programming_languages,id'])->passes()) {
-            if(!$user->programmingLanguages->contains($programming_language_id)) {
-                $user->programmingLanguages()->attach($programming_language_id);
-                $user->load('programmingLanguages'); //Refreshes the model
+        if ($user->id == $request->user()->id) {
+            if(\Validator::make(['programming_language_id' => $programming_language_id], ['programming_language_id' => 'exists:programming_languages,id'])->passes()) {
+                if(!$user->programmingLanguages->contains($programming_language_id)) {
+                    $user->programmingLanguages()->attach($programming_language_id);
+                    $user->load('programmingLanguages'); //Refreshes the model
+                }
+                return new CustomJsonResponse(true, $user, 200);
             }
-            return new CustomJsonResponse(true, $user, 200);
+            return new CustomJsonResponse(false, "Programming Language does not exist", 404);
+        } else {
+            return new CustomJsonResponse(false, "No permission to edit member info", 403);
         }
-        return new CustomJsonResponse(false, "Programming Language does not exist", 404);
+
     }
 
     public function deleteProgrammingLanguage(Request $request, $identification, $programming_language_id) {
@@ -285,13 +326,18 @@ class UsersController extends Controller
             return new CustomJsonResponse(false,"User not found", 404);
         }
 
-        foreach ($user->programmingLanguages as $programmingLanguage) {
-            if($programmingLanguage->id == $programming_language_id) {
-                $user->programmingLanguages()->detach($programming_language_id);
-                $user->load('programmingLanguages'); //Refreshes the model
-                return new CustomJsonResponse(true, $user, 200);
+        if ($user->id == $request->user()->id) {
+            foreach ($user->programmingLanguages as $programmingLanguage) {
+                if($programmingLanguage->id == $programming_language_id) {
+                    $user->programmingLanguages()->detach($programming_language_id);
+                    $user->load('programmingLanguages'); //Refreshes the model
+                    return new CustomJsonResponse(true, $user, 200);
+                }
             }
+            return new CustomJsonResponse(false, "User doesn't know this programming language", 404);
+        } else {
+            return new CustomJsonResponse(false, "No permission to edit member info", 403);
         }
-        return new CustomJsonResponse(false, "User doesn't know this programming language", 404);
+
     }
 }
